Final Version

+++ Task.h
@@ -12,50 +12,59 @@ typedef struct tcb{
 } Task;
 
 
+// Initialize task t
 void TaskInit(Task* t, int pid, int prio, char* name){
 	t->pid = pid;
 	t->prio = prio;
 	strcpy(t->name, name);
-};
+};	
 
-
+//  Get task name
 char* taskGetName(Task* task){
 	return task->name;
 };
 
-unsigned long taskGetId(Task* task){
+// Get task ID
+int taskGetId(Task* task){
 	return task->pid;
 };
+
 
 // bool taskShouldSuspend(Task* task){
 // 
 // };
 
+
 // void taskSuspend(Task* task){
 // 	
 // };
 
+// Delay execution by tick time 
 void taskWait(int t_sleep){
 	sleep(t_sleep);
 };
 
+// Ask for access to memory
 void taskGetMem(struct Mutex* m){
 	MutexAcquire(m);
 };
 
+// Replace memory by new value
 void taskSetMem(char* memory, char* new_memory){
 	strcpy(memory, new_memory);
 };
 
+// Release the access to memory
 void taskReleaseMem(struct Mutex* m){
 	MutexRelease(m);
 };
 
-unsigned int taskPrio(Task* task){
+// Get task priority
+int taskPrio(Task* task){
 	return task->prio;
 };
 
-
+// Struct to pass arguments for taskWake function (by pthread_create)
 typedef struct ScheduleInfo_t{
                 Task* task;
                 char* memory;
@@ -63,6 +72,7 @@ typedef struct ScheduleInfo_t{
                 int tick_time;
     } ScheduleInfo;
 
+// main tasks execution body
 void* taskWake(void* args){
 
 	ScheduleInfo* sc_info = (ScheduleInfo*)args;
